remove old FuelRecipeLogic and FueledMultiblockController
